Can select, sign and un-sign players.
The unsign button isnt functional but it does work.
There is no checks done to see if a player is already on a team so you can add multiple of the same person to a team.
You can also sign people who are already on a team.
Active team layout is also incomplete.