v1.2 - Fixed Stages
This version is major as everything was rebuilt after the IDE broke the project and messed up all the files for an unknown reason. Everything is up and working again. Some changes include:
New mouse pointer when hovering over the active team players
Fixed stages so that swap team and active team management could not be viewed simultaneously.